requirements.txt
google-cloud-storage
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
requests
